# apps/ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Customer Service Chatbot API",
    "description": "AI-powered customer service chatbot with LangGraph agentic workflow",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/chat/sessions": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Create Chat Session",
        "description": "Create a new chat session for the authenticated user.\n\nArgs:\n    request: The session creation parameters.\n    current_user: The authenticated user.\n    database_session: The async database session.\n\nReturns:\n    The newly created chat session.",
        "operationId": "create_chat_session_api_chat_sessions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSessionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "List Chat Sessions",
        "description": "List all chat sessions for the authenticated user.\n\nArgs:\n    include_archived: Whether to include archived sessions.\n    current_user: The authenticated user.\n    database_session: The async database session.\n\nReturns:\n    A list of session summaries with message counts.",
        "operationId": "list_chat_sessions_api_chat_sessions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "include_archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Archived"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatSessionSummary"
                  },
                  "title": "Response List Chat Sessions Api Chat Sessions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/sessions/{session_id}": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Chat Session",
        "description": "Retrieve a specific chat session with all messages.\n\nArgs:\n    session_id: The ID of the session to retrieve.\n    current_user: The authenticated user.\n    database_session: The async database session.\n\nReturns:\n    The full chat session with its message history.\n\nRaises:\n    HTTPException: If the session is not found.",
        "operationId": "get_chat_session_api_chat_sessions__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "chat"
        ],
        "summary": "Update Chat Session",
        "description": "Update a chat session's title or archive status.\n\nArgs:\n    session_id: The ID of the session to update.\n    request: The fields to update.\n    current_user: The authenticated user.\n    database_session: The async database session.\n\nReturns:\n    The updated chat session.\n\nRaises:\n    HTTPException: If the session is not found.",
        "operationId": "update_chat_session_api_chat_sessions__session_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSessionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "chat"
        ],
        "summary": "Delete Chat Session",
        "description": "Permanently delete a chat session and all its messages.\n\nArgs:\n    session_id: The ID of the session to delete.\n    current_user: The authenticated user.\n    database_session: The async database session.\n\nRaises:\n    HTTPException: If the session is not found.",
        "operationId": "delete_chat_session_api_chat_sessions__session_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/sessions/{session_id}/messages": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Send Message",
        "description": "Send a user message and receive an AI-generated response.\n\nThis endpoint persists the user message, retrieves short-term memory\nfrom mem0, runs the LangGraph agentic workflow, persists the assistant\nresponse, and updates mem0 with the new exchange.\n\nArgs:\n    session_id: The ID of the session to send the message in.\n    request: The user's message content.\n    current_user: The authenticated user.\n    database_session: The async database session.\n\nReturns:\n    The assistant's response with optional agent trace data.\n\nRaises:\n    HTTPException: If the session is not found.",
        "operationId": "send_message_api_chat_sessions__session_id__messages_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/sessions/{session_id}/messages/stream": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Send Message Stream",
        "description": "Stream a user message response with intermediate thinking steps via SSE.\n\nEmits events as the workflow progresses through planner, executor (per agent),\nand synthesizer stages, then a final 'done' event with the persisted message.\n\nArgs:\n    session_id: The ID of the session to send the message in.\n    request: The user's message content.\n    current_user: The authenticated user.\n    database_session: The async database session.\n\nReturns:\n    A streaming response with text/event-stream content type.\n\nRaises:\n    HTTPException: If the session is not found.",
        "operationId": "send_message_stream_api_chat_sessions__session_id__messages_stream_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ChatCompletionResponse": {
        "properties": {
          "message": {
            "$ref": "#/components/schemas/ChatMessageResponse"
          },
          "agent_trace": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Trace",
            "description": "Optional trace data from the agentic workflow execution"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatCompletionResponse",
        "description": "The assistant's response after processing a user message."
      },
      "ChatMessageRequest": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Content",
            "description": "The user's message text"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ChatMessageRequest",
        "description": "A user message sent to the chatbot for processing."
      },
      "ChatMessageResponse": {
        "properties": {
          "message_id": {
            "type": "string",
            "format": "uuid",
            "title": "Message Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "message_id",
          "session_id",
          "role",
          "content",
          "created_at"
        ],
        "title": "ChatMessageResponse",
        "description": "A single message within a chat conversation."
      },
      "ChatSessionCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 500,
            "title": "Title",
            "default": "New Chat"
          }
        },
        "type": "object",
        "title": "ChatSessionCreate",
        "description": "Request body for creating a new chat session."
      },
      "ChatSessionResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "is_archived": {
            "type": "boolean",
            "title": "Is Archived"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageResponse"
            },
            "type": "array",
            "title": "Messages",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "user_id",
          "title",
          "is_archived",
          "created_at",
          "updated_at"
        ],
        "title": "ChatSessionResponse",
        "description": "Full representation of a chat session with its messages."
      },
      "ChatSessionSummary": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "is_archived": {
            "type": "boolean",
            "title": "Is Archived"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "message_count": {
            "type": "integer",
            "title": "Message Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "title",
          "is_archived",
          "created_at",
          "updated_at"
        ],
        "title": "ChatSessionSummary",
        "description": "Lightweight session representation for listing endpoints."
      },
      "ChatSessionUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "is_archived": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Archived"
          }
        },
        "type": "object",
        "title": "ChatSessionUpdate",
        "description": "Request body for updating a chat session's metadata."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}